+++ gradio/layouts/tabs.py
@@ -23,7 +23,9 @@ class Tabs(BlockContext, metaclass=ComponentMeta):
         visible: bool = True,
         elem_id: str | None = None,
         elem_classes: list[str] | str | None = None,
-        **kwargs,
+        render: bool = True,
+        root_url: str | None = None,
+        _skip_init_processing: bool = False,
     ):
         """
         Parameters:
@@ -31,20 +33,19 @@ class Tabs(BlockContext, metaclass=ComponentMeta):
             visible: If False, Tabs will be hidden.
             elem_id: An optional string that is assigned as the id of this component in the HTML DOM. Can be used for targeting CSS styles.
             elem_classes: An optional string or list of strings that are assigned as the class of this component in the HTML DOM. Can be used for targeting CSS styles.
+            render: If False, this layout will not be rendered in the Blocks context. Should be used if the intention is to assign event listeners now but render the component later.
+            root_url: The remote URL that of the Gradio app that this layout belongs to. Used in `gr.load()`. Should not be set manually.
         """
         BlockContext.__init__(
-            self, visible=visible, elem_id=elem_id, elem_classes=elem_classes, **kwargs
+            self,
+            visible=visible,
+            elem_id=elem_id,
+            elem_classes=elem_classes,
+            render=render,
+            root_url=root_url,
+            _skip_init_processing=_skip_init_processing,
         )
         self.selected = selected
-
-    @staticmethod
-    def update(
-        selected: int | str | None = None,
-    ):
-        return {
-            "selected": selected,
-            "__type__": "update",
-        }
 
 
 @document()
@@ -66,12 +67,14 @@ class Tab(BlockContext, metaclass=ComponentMeta):
 
     def __init__(
         self,
-        label: str,
+        label: str | None = None,
         *,
         id: int | str | None = None,
         elem_id: str | None = None,
         elem_classes: list[str] | str | None = None,
-        **kwargs,
+        render: bool = True,
+        root_url: str | None = None,
+        _skip_init_processing: bool = False,
     ):
         """
         Parameters:
@@ -81,7 +84,12 @@ class Tab(BlockContext, metaclass=ComponentMeta):
             elem_classes: An optional string or list of strings that are assigned as the class of this component in the HTML DOM. Can be used for targeting CSS styles.
         """
         BlockContext.__init__(
-            self, elem_id=elem_id, elem_classes=elem_classes, **kwargs
+            self,
+            elem_id=elem_id,
+            elem_classes=elem_classes,
+            render=render,
+            root_url=root_url,
+            _skip_init_processing=_skip_init_processing,
         )
         self.label = label
         self.id = id