+++ gradio/routes.py
@@ -16,6 +16,7 @@ import mimetypes
 import os
 import posixpath
 import secrets
+import shutil
 import tempfile
 import threading
 import time
@@ -25,11 +26,12 @@ from pathlib import Path
 from queue import Empty as EmptyQueue
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type
 
+import anyio
 import fastapi
 import httpx
 import markupsafe
 import orjson
-from fastapi import Depends, FastAPI, File, HTTPException, UploadFile, WebSocket, status
+from fastapi import Depends, FastAPI, HTTPException, WebSocket, status
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import (
     FileResponse,
@@ -39,25 +41,27 @@ from fastapi.responses import (
 )
 from fastapi.security import OAuth2PasswordRequestForm
 from fastapi.templating import Jinja2Templates
+from gradio_client import utils as client_utils
 from gradio_client.documentation import document, set_documentation_group
 from jinja2.exceptions import TemplateNotFound
+from multipart.multipart import parse_options_header
 from starlette.background import BackgroundTask
 from starlette.responses import RedirectResponse, StreamingResponse
 
 import gradio
 import gradio.ranged_response as ranged_response
-from gradio import processing_utils, route_utils, utils, wasm_utils
+from gradio import route_utils, utils, wasm_utils
 from gradio.context import Context
-from gradio.data_classes import (
-    ComponentServerBody,
-    PredictBody,
-    ResetBody,
-)
-from gradio.deprecation import warn_deprecation
+from gradio.data_classes import ComponentServerBody, PredictBody, ResetBody
 from gradio.exceptions import Error
 from gradio.oauth import attach_oauth
 from gradio.queueing import Estimation, Event
-from gradio.route_utils import Request  # noqa: F401
+from gradio.route_utils import (  # noqa: F401
+    GradioMultiPartParser,
+    GradioUploadFile,
+    MultiPartException,
+    Request,
+)
 from gradio.state_holder import StateHolder
 from gradio.utils import (
     cancel_tasks,
@@ -668,16 +672,42 @@ class App(FastAPI):
             return app.get_blocks()._queue.get_estimation()
 
         @app.post("/upload", dependencies=[Depends(login_check)])
-        async def upload_file(
-            files: List[UploadFile] = File(...),
-        ):
-            output_files = []
-            for input_file in files:
-                output_files.append(
-                    await processing_utils.save_uploaded_file(
-                        input_file, app.uploaded_file_dir, app.get_blocks().limiter
-                    )
+        async def upload_file(request: fastapi.Request):
+            content_type_header = request.headers.get("Content-Type")
+            content_type: bytes
+            content_type, _ = parse_options_header(content_type_header)
+            if content_type != b"multipart/form-data":
+                raise HTTPException(status_code=400, detail="Invalid content type.")
+
+            try:
+                multipart_parser = GradioMultiPartParser(
+                    request.headers,
+                    request.stream(),
+                    max_files=1000,
+                    max_fields=1000,
                 )
+                form = await multipart_parser.parse()
+            except MultiPartException as exc:
+                raise HTTPException(status_code=400, detail=exc.message) from exc
+
+            output_files = []
+            for temp_file in form.getlist("files"):
+                assert isinstance(temp_file, GradioUploadFile)
+                if temp_file.filename:
+                    file_name = Path(temp_file.filename).name
+                    name = client_utils.strip_invalid_filename_characters(file_name)
+                else:
+                    name = f"tmp{secrets.token_hex(5)}"
+                directory = Path(app.uploaded_file_dir) / temp_file.sha.hexdigest()
+                directory.mkdir(exist_ok=True, parents=True)
+                dest = (directory / name).resolve()
+                await anyio.to_thread.run_sync(
+                    shutil.move,
+                    temp_file.file.name,
+                    dest,
+                    limiter=app.get_blocks().limiter,
+                )
+                output_files.append(dest)
             return output_files
 
         @app.on_event("startup")
@@ -756,7 +786,6 @@ def mount_gradio_app(
     app: fastapi.FastAPI,
     blocks: gradio.Blocks,
     path: str,
-    gradio_api_url: str | None = None,
     app_kwargs: dict[str, Any] | None = None,
 ) -> fastapi.FastAPI:
     """Mount a gradio.Blocks to an existing FastAPI application.
@@ -765,7 +794,6 @@ def mount_gradio_app(
         app: The parent FastAPI application.
         blocks: The blocks object we want to mount to the parent app.
         path: The path at which the gradio application will be mounted.
-        gradio_api_url: Deprecated and has no effect.
         app_kwargs: Additional keyword arguments to pass to the underlying FastAPI app as a dictionary of parameter keys and argument values. For example, `{"docs_url": "/docs"}`
     Example:
         from fastapi import FastAPI
@@ -782,9 +810,6 @@ def mount_gradio_app(
     blocks.config = blocks.get_config_file()
     blocks.validate_queue_settings()
     gradio_app = App.create_app(blocks, app_kwargs=app_kwargs)
-
-    if gradio_api_url is not None:
-        warn_deprecation("gradio_api_url is deprecated and has not effect.")
 
     @app.on_event("startup")
     async def start_queue():