+++ js/_website/src/routes/[[version]]/docs/themes/+page.server.ts
@@ -14,7 +14,16 @@ const COLOR_SETS = [
 ];
 
 export async function load({ parent }) {
-	const { components, helpers, modals, py_client, routes, docs, on_main, wheel } = await parent();
+	const {
+		components,
+		helpers,
+		modals,
+		py_client,
+		routes,
+		docs,
+		on_main,
+		wheel
+	} = await parent();
 
 	let objs = [docs.building.base];
 	let headers = [
@@ -88,7 +97,7 @@ export async function load({ parent }) {
 		COLOR_SETS,
 		headers,
 		method_headers,
-		on_main, 
+		on_main,
 		wheel
 	};
 }