+++ test/test_gradio_component_cli.py
@@ -4,6 +4,7 @@ import pytest
 
 from gradio.cli.commands.components._create_utils import OVERRIDES
 from gradio.cli.commands.components.create import _create
+from gradio.cli.commands.components.show import _show
 
 
 @pytest.mark.parametrize(
@@ -56,13 +57,6 @@ def test_raise_error_component_template_does_not_exist(tmp_path):
         )
 
 
-def test_overwrite_deletes_previous_content(tmp_path):
-    _create("MyGallery", tmp_path, template="Gallery", overwrite=True)
-    _create("MySlider", tmp_path, template="Slider", overwrite=True)
-    assert (tmp_path / "frontend" / "interactive" / "InteractiveSlider.svelte").exists()
-    assert not (tmp_path / "frontend" / "static" / "StaticGallery.svelte").exists()
-
-
 def test_do_not_replace_class_name_in_import_statement(tmp_path):
     _create("MyImage", template="Image", directory=tmp_path, overwrite=True)
     code = (tmp_path / "backend" / "gradio_myimage" / "myimage.py").read_text()
@@ -76,3 +70,13 @@ def test_raises_if_directory_exists(tmp_path):
         ValueError, match=f"The directory {tmp_path.resolve()} already exists."
     ):
         _create("MyComponent", tmp_path)
+
+
+def test_show(capsys):
+    _show()
+    stdout, _ = capsys.readouterr()
+    assert "Form Component" in stdout
+    assert "Beginner Friendly" in stdout
+    assert "Layout" in stdout
+    assert "Dataframe" not in stdout
+    assert "Dataset" not in stdout